+++ backend/ddang/src/main/java/com/ddang/ddang/chat/application/ChatRoomService.java
@@ -91,7 +91,7 @@ public class ChatRoomService {
     public ReadParticipatingChatRoomDto readByChatRoomId(final Long chatRoomId, final Long userId) {
         final User findUser = userRepository.findById(userId)
                                             .orElseThrow(() -> new UserNotFoundException("사용자 정보를 찾을 수 없습니다."));
-        final ChatRoom chatRoom = chatRoomRepository.findById(chatRoomId)
+        final ChatRoom chatRoom = chatRoomRepository.findChatRoomById(chatRoomId)
                                                     .orElseThrow(() ->
                                                             new ChatRoomNotFoundException(
                                                                     "지정한 아이디에 대한 채팅방을 찾을 수 없습니다."
