+++ backend/ddang/src/main/java/com/ddang/ddang/chat/infrastructure/persistence/QuerydslChatRoomRepository.java
@@ -9,5 +9,7 @@ public interface QuerydslChatRoomRepository {
 
     List<ChatRoom> findAllByUserId(final Long userId);
 
+    Optional<ChatRoom> findChatRoomById(final Long chatRoomId);
+
     Optional<ChatRoom> findByAuctionId(final Long auctionId);
 }
