+++ backend/ddang/src/test/java/com/ddang/ddang/chat/infrastructure/persistence/JpaChatRoomRepositoryTest.java
@@ -4,8 +4,6 @@ import com.ddang.ddang.auction.domain.Auction;
 import com.ddang.ddang.auction.domain.BidUnit;
 import com.ddang.ddang.auction.domain.Price;
 import com.ddang.ddang.auction.infrastructure.persistence.JpaAuctionRepository;
-import com.ddang.ddang.bid.domain.Bid;
-import com.ddang.ddang.bid.domain.BidPrice;
 import com.ddang.ddang.bid.infrastructure.persistence.JpaBidRepository;
 import com.ddang.ddang.category.domain.Category;
 import com.ddang.ddang.category.infrastructure.persistence.JpaCategoryRepository;
@@ -25,7 +23,6 @@ import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.context.annotation.Import;
 
 import java.time.LocalDateTime;
-import java.util.List;
 import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -114,146 +111,6 @@ class JpaChatRoomRepositoryTest {
             softAssertions.assertThat(actual.get().getId()).isEqualTo(expected.getId());
             softAssertions.assertThat(actual.get().getAuction()).isEqualTo(expected.getAuction());
             softAssertions.assertThat(actual.get().getBuyer()).isEqualTo(expected.getBuyer());
-        });
-    }
-
-    @Test
-    void 지정한_사용자_아이디가_포함된_채팅방을_조회한다() {
-        // given
-        final Category main = new Category("메인");
-        final Category sub = new Category("서브");
-        main.addSubCategory(sub);
-        categoryRepository.save(main);
-
-
-        final User merry = User.builder()
-                               .name("메리")
-                               .profileImage("profile.png")
-                               .reliability(4.7d)
-                               .oauthId("12345")
-                               .build();
-        final User encho = User.builder()
-                               .name("엔초")
-                               .profileImage("profile.png")
-                               .reliability(4.7d)
-                               .oauthId("12346")
-                               .build();
-        final User jamie = User.builder()
-                               .name("제이미")
-                               .profileImage("profile.png")
-                               .reliability(4.7d)
-                               .oauthId("12347")
-                               .build();
-        final User zeeto = User.builder()
-                               .name("지토")
-                               .profileImage("profile.png")
-                               .reliability(4.7d)
-                               .oauthId("12348")
-                               .build();
-
-        userRepository.save(merry);
-        userRepository.save(encho);
-        userRepository.save(jamie);
-        userRepository.save(zeeto);
-
-        final Auction merryAuction = Auction.builder()
-                                            .title("경매 1")
-                                            .seller(merry)
-                                            .subCategory(sub)
-                                            .bidUnit(new BidUnit(1_000))
-                                            .startPrice(new Price(10_000))
-                                            .build();
-        final Auction enchoAuction = Auction.builder()
-                                            .title("경매 2")
-                                            .seller(encho)
-                                            .subCategory(sub)
-                                            .bidUnit(new BidUnit(2_000))
-                                            .startPrice(new Price(20_000))
-                                            .build();
-        final Auction jamieAuction = Auction.builder()
-                                            .title("경매 3")
-                                            .seller(jamie)
-                                            .subCategory(sub)
-                                            .bidUnit(new BidUnit(3_000))
-                                            .startPrice(new Price(30_000))
-                                            .build();
-
-        auctionRepository.save(merryAuction);
-        auctionRepository.save(enchoAuction);
-        auctionRepository.save(jamieAuction);
-
-        final ChatRoom merryZeeto = new ChatRoom(merryAuction, zeeto);
-        final ChatRoom enchoZeeto = new ChatRoom(enchoAuction, zeeto);
-        final ChatRoom jamieEncho = new ChatRoom(jamieAuction, encho);
-        chatRoomRepository.save(merryZeeto);
-        chatRoomRepository.save(enchoZeeto);
-        chatRoomRepository.save(jamieEncho);
-
-        em.flush();
-        em.clear();
-
-        // when
-        final List<ChatRoom> actual = chatRoomRepository.findAllByUserId(encho.getId());
-
-        // then
-        SoftAssertions.assertSoftly(softAssertions -> {
-            softAssertions.assertThat(actual).hasSize(2);
-            softAssertions.assertThat(actual.get(1).getId()).isEqualTo(enchoZeeto.getId());
-            softAssertions.assertThat(actual.get(0).getId()).isEqualTo(jamieEncho.getId());
-        });
-    }
-
-    @Test
-    void 지정한_경매_아이디가_포함된_채팅방을_조회한다() {
-        // given
-        final Category main = new Category("메인");
-        final Category sub = new Category("서브");
-        main.addSubCategory(sub);
-        categoryRepository.save(main);
-
-        final User seller = User.builder()
-                                .name("회원1")
-                                .profileImage("profile.png")
-                                .reliability(4.7d)
-                                .oauthId("12345")
-                                .build();
-        final User buyer = User.builder()
-                               .name("회원2")
-                               .profileImage("profile.png")
-                               .reliability(4.7d)
-                               .oauthId("12346")
-                               .build();
-        userRepository.save(seller);
-        userRepository.save(buyer);
-
-        final Auction auction = Auction.builder()
-                                       .title("경매 1")
-                                       .seller(seller)
-                                       .subCategory(sub)
-                                       .bidUnit(new BidUnit(1_000))
-                                       .startPrice(new Price(10_000))
-                                       .closingTime(LocalDateTime.now().minusDays(3L))
-                                       .build();
-        auctionRepository.save(auction);
-
-        final Bid bid = new Bid(auction, buyer, new BidPrice(15_000));
-        bidRepository.save(bid);
-
-        auction.updateLastBid(bid);
-
-        final ChatRoom chatRoom = new ChatRoom(auction, buyer);
-        chatRoomRepository.save(chatRoom);
-
-        em.flush();
-        em.clear();
-
-        // when
-        final Optional<ChatRoom> actual = chatRoomRepository.findByAuctionId(auction.getId());
-
-        // then
-        SoftAssertions.assertSoftly(softAssertions -> {
-            softAssertions.assertThat(actual).isPresent();
-            softAssertions.assertThat(actual.get()).isEqualTo(chatRoom);
         });
     }
